% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BipolarMisclassification.R
\docType{package}
\name{BipolarMisclassification}
\alias{BipolarMisclassification}
\alias{BipolarMisclassification-package}
\title{BipolarMisclassification}
\description{
BipolarMisclassification
}
